Replace set with expose_to_application
Middleman replaced set with config but this was actually being used to
explose things on the middleman application.

I don't think that either of these are needed outside of this
extenssion, but to avoid further investigation and changes I'm just
going to put them out on `app` for now. They're not going to colide with
anyone else, they just don't need to be exposed and should be hidden in
the future.