Work with multisite middleman-blog
My guess is that when middleman-blog moved to multisite support they
stopped hanging `blog` on the main application. Accessing blog data is a
little cumbersome now, but internally middleman-blog actually does
something similar (if you don't specify a blog they assume the first).

For now we're hard coded to the first blog. A better way might be to
consolidate ALL of the blog authors.